ADM-368 rename exportStep to reportStep
