+++ frontend/src/components/Metrics/ReportStep/index.tsx
@@ -1,18 +1,21 @@
-import { Velocity } from '@src/components/Metrics/ExportStep/Velocity'
+import { Velocity } from '@src/components/Metrics/ReportStep/Velocity'
 import { useEffect, useState } from 'react'
 import { useGeneratorReportEffect } from '@src/hooks/useGeneratorReportEffect'
 import { Loading } from '@src/components/Loading'
 
-export const ExportStep = () => {
+export const ReportStep = () => {
   const { generatorReport, isLoading } = useGeneratorReportEffect()
-  const [velocityData, setVelocityData] = useState({ velocityForSP: '2', velocityForCards: '2' })
+  const [velocityData, setVelocityData] = useState({
+    velocityForSP: '2',
+    velocityForCards: '2',
+  })
   useEffect(() => {
     generatorReport().then((res) => {
       if (res) {
         setVelocityData(res.response?.velocity)
       }
     })
-  }, [generatorReport])
+  }, [])
   return (
     <>
       {isLoading && <Loading />}
